+++ src/main/java/com/github/jlangch/aviron/ex/QuarantineException.java
@@ -1,0 +1,47 @@
+/*                 _                 
+ *       /\       (_)            
+ *      /  \__   ___ _ __ ___  _ __  
+ *     / /\ \ \ / / | '__/ _ \| '_ \ 
+ *    / ____ \ V /| | | | (_) | | | |
+ *   /_/    \_\_/ |_|_|  \___/|_| |_|
+ *
+ *
+ * Copyright 2025 Aviron
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.jlangch.aviron.ex;
+
+
+/**
+ * Exception thrown if on severe quarantine error
+ * 
+ * @author juerg
+ */
+public class QuarantineException extends RuntimeException {
+
+    public QuarantineException(final String message) {
+        super(message);
+    }
+
+    public QuarantineException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+
+    public QuarantineException(final Throwable cause) {
+        super(cause);
+    }
+
+    
+    private static final long serialVersionUID = 1287052357875193031L;
+}